Move `PlayerPokemon` to `src/field/player-pokemon.ts`
