+++ src/@types/PokemonModifierTransferSelectFilter.ts
@@ -1,4 +1,4 @@
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import type { PokemonHeldItemModifier } from "#app/modifier/modifier";
 
 export type PokemonModifierTransferSelectFilter = (
