+++ src/@types/PokemonSelectFilter.ts
@@ -1,3 +1,3 @@
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 
 export type PokemonSelectFilter = (pokemon: PlayerPokemon) => string | null;
