+++ src/battle-scene.ts
@@ -41,7 +41,8 @@ import { NewArenaEvent } from "#app/events/battle-scene";
 import { Arena, ArenaBase } from "#app/field/arena";
 import DamageNumberHandler from "#app/field/damage-number-handler";
 import { EnemyPokemon } from "#app/field/enemy-pokemon";
-import { PlayerPokemon, type Pokemon } from "#app/field/pokemon";
+import { PlayerPokemon } from "#app/field/player-pokemon";
+import type { Pokemon } from "#app/field/pokemon";
 import PokemonSpriteSparkleHandler from "#app/field/pokemon-sprite-sparkle-handler";
 import Trainer from "#app/field/trainer";
 import { type GameMode, getGameMode } from "#app/game-mode";
