+++ src/data/daily-run.ts
@@ -1,6 +1,6 @@
 import type { SpeciesId } from "#enums/species-id";
 import { globalScene } from "#app/global-scene";
-import { PlayerPokemon } from "#app/field/pokemon";
+import { PlayerPokemon } from "#app/field/player-pokemon";
 import type { StarterConfig } from "#app/@types/StarterConfig";
 import { randSeedGauss, randSeedInt, randSeedItem } from "#app/utils";
 import type { PokemonSpeciesForm } from "./pokemon-species-form";
