+++ src/data/egg-hatch-data.ts
@@ -1,5 +1,5 @@
 import { globalScene } from "#app/global-scene";
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import type { DexEntry } from "#app/@types/DexData";
 import type { StarterDataEntry } from "#app/@types/StarterData";
