+++ src/data/mystery-encounters/encounters/berries-abound-encounter.ts
@@ -8,7 +8,7 @@ import {
   setEncounterExp,
   setEncounterRewards,
 } from "#app/data/mystery-encounters/utils/encounter-phase-utils";
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import type { Pokemon } from "#app/field/pokemon";
 import { EnemyPokemon } from "#app/field/enemy-pokemon";
 import type { BerryModifierType, ModifierTypeOption } from "#app/modifier/modifier-type";
