+++ src/data/mystery-encounters/encounters/field-trip-encounter.ts
@@ -7,7 +7,7 @@ import {
   setEncounterExp,
   setEncounterRewards,
 } from "#app/data/mystery-encounters/utils/encounter-phase-utils";
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import type { PokemonMove } from "#app/field/pokemon-move";
 import { modifierTypes } from "#app/modifier/modifier-types";
 import type { OptionSelectItem } from "#app/ui/interfaces/option-select-config";
