+++ src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -13,7 +13,8 @@ import type PokemonSpecies from "#app/data/pokemon-species";
 import { Status } from "#app/data/status-effect";
 import type { TrainerConfig } from "#app/data/trainer-config";
 import type { Variant } from "#app/data/variant";
-import { type PlayerPokemon, type Pokemon } from "#app/field/pokemon";
+import { type Pokemon } from "#app/field/pokemon";
+import { type PlayerPokemon } from "#app/field/player-pokemon";
 import { PokemonMove } from "#app/field/pokemon-move";
 import { PokemonSummonData } from "#app/field/pokemon-summon-data";
 import Trainer from "#app/field/trainer";
