+++ src/phases/abstract-player-party-member-pokemon-phase.ts
@@ -1,4 +1,4 @@
-import type { PlayerPokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import { PartyMemberPokemonPhase } from "./abstract-party-member-pokemon-phase";
 
 // TODO: Delete this phase and replace it with PokemonPhase
