+++ src/phases/evolution-phase.ts
@@ -6,7 +6,8 @@ import { type FormChangePhase } from "#app/phases/form-change-phase";
 import type { AnySound } from "#app/audio-manager";
 import { EVOLVE_MOVE } from "#app/data/pokemon-level-moves";
 import type { SpeciesFormEvolution } from "#app/data/pokemon-evolutions";
-import type { PlayerPokemon, Pokemon } from "#app/field/pokemon";
+import type { Pokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import { globalScene } from "#app/global-scene";
 import { getPokemonNameWithAffix } from "#app/messages";
 import { EndEvolutionPhase } from "#app/phases/end-evolution-phase";
