+++ src/ui/handlers/party-ui-handler.ts
@@ -9,7 +9,8 @@ import { getGenderSymbol, getGenderTextStyle } from "#app/data/gender";
 import { ForceSwitchOutAttr } from "#app/data/moves/move-attrs/force-switch-out-attr";
 import { SpeciesFormChangeItemTrigger } from "#app/data/species-form-change-triggers/species-form-change-item-trigger";
 import { getVariantTint } from "#app/data/variant";
-import type { PlayerPokemon, Pokemon } from "#app/field/pokemon";
+import type { Pokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import { globalScene } from "#app/global-scene";
 import { getPokemonNameWithAffix } from "#app/messages";
 import { type PokemonFormChangeItemModifier, type PokemonHeldItemModifier } from "#app/modifier/modifier";
