+++ test/test-utils/gameManager.ts
@@ -7,7 +7,8 @@ import { type TurnEndPhase } from "#app/phases/turn-end-phase";
 
 import { updateUserInfo } from "#app/account";
 import BattleScene from "#app/battle-scene";
-import type { PlayerPokemon, Pokemon } from "#app/field/pokemon";
+import type { Pokemon } from "#app/field/pokemon";
+import type { PlayerPokemon } from "#app/field/player-pokemon";
 import type { EnemyPokemon } from "#app/field/enemy-pokemon";
 import Trainer from "#app/field/trainer";
 import { getGameMode } from "#app/game-mode";
